+++ src/stock_str.rs
@@ -8,7 +8,7 @@ use strum::EnumProperty;
 use strum_macros::EnumProperty;
 
 use crate::blob::BlobObject;
-use crate::chat::{self, ChatId, ProtectionStatus};
+use crate::chat::{self, Chat, ChatId, ProtectionStatus};
 use crate::config::Config;
 use crate::constants::{Viewtype, DC_CONTACT_ID_SELF};
 use crate::contact::{Contact, Origin};
@@ -330,6 +330,12 @@ pub enum StockMessage {
 
     #[strum(props(fallback = "%1$s replied, waiting for being added to the group…"))]
     SecureJoinReplies = 118,
+
+    #[strum(props(fallback = "Scan to chat with %1$s"))]
+    SetupContactQRDescription = 119,
+
+    #[strum(props(fallback = "Scan to join group %1$s"))]
+    SecureJoinGroupQRDescription = 120,
 }
 
 impl StockMessage {
@@ -612,6 +618,29 @@ pub(crate) async fn secure_join_replies(context: &Context, contact_id: u32) -> S
     } else {
         format!("secure_join_replies: unknown contact {}", contact_id)
     }
+}
+
+/// Stock string: `Scan to chat with %1$s`.
+pub(crate) async fn setup_contact_qr_description(
+    context: &Context,
+    display_name: &str,
+    addr: &str,
+) -> String {
+    let name = if display_name == addr {
+        addr.to_owned()
+    } else {
+        format!("{} ({})", display_name, addr)
+    };
+    translated(context, StockMessage::SetupContactQRDescription)
+        .await
+        .replace1(name)
+}
+
+/// Stock string: `Scan to join %1$s`.
+pub(crate) async fn secure_join_group_qr_description(context: &Context, chat: &Chat) -> String {
+    translated(context, StockMessage::SecureJoinGroupQRDescription)
+        .await
+        .replace1(chat.get_name())
 }
 
 /// Stock string: `%1$s verified.`.